requirements.txt
# Overall project requirements
# This file lists dependencies needed for development and testing

# Backend services
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
redis==5.0.1
python-multipart==0.0.6
pandas==2.1.3
numpy==1.24.3
faiss-cpu==1.7.4
pytest==7.4.3
httpx==0.25.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
google-generativeai==0.3.1

# Frontend development
# (No Python dependencies needed for HTML/CSS/JS frontend)

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.5